Create 5b.py
Write a python class to reverse a sentence (initialized via constructor) word by word. That
is: “I am here” should be reversed as “here am I”. Create instances of this class for each
of the three strings input by the user and display the reversed string for each, in
descending order of number of vowels in the string.